+++ package.json
@@ -3,7 +3,10 @@
     "type": "module",
     "version": "1.0.0",
     "description": "",
-    "main": "app.js",
+    "main": "index.js",
+    "engines": {
+        "node": "16.x"
+    },
     "dependencies": {
         "@notionhq/client": "^0.1.9",
         "Math": "^0.0.1-security",
